Kosmostation Map
Initial of the Kosmostation Map. Currently only features Bridge - not completed fully. The bridge contains (of course) a hammer and sickle, red plastitanium floor that blends in with the hammer and sickle, and some computer monitors, with security records and security cameras. Station alerts, for both air and other alarms. This type of bridge also features a special desk with DENIED and GRANTED stamps - which are mostly there for roleplaying and environment  building, but may also have a mechanical use in the future, giving command more things to do such as approving certain things or rejecting them, which may include parole, accepting certain deliveries, and such.

TODO:
- Add a medical records console
- Decorate the map a bit further
- Add the rest of the station, duh
- Achieve true communism